Pass the source file name to the transpiler
When transpiling files as modules, the filename is passed to TypeScript so it can interpret it properly. This is necessary to respect the behavior of the command line, especially when deciding whether files should be interpreted as jsx/tsx files, as per [the transpiler reference](https://github.com/Microsoft/TypeScript/wiki/Using-the-Compiler-API#transpiling-a-single-file).

See #213.