Fix plugins field type from FuseBoxOptions
The old type was `Plugin[] | [Plugin[]]` ((array of Plugin) **or** (array of array of Plugin)) while I suggest `Array<Plugin | Plugin[]>`  (array of (Plugin **or** array of Plugin)). Only the latest is compatible with `fuse-box-examples` which have, on many occasions, an array of CSS-related plugins mixed up with single plugins.