Update WebIndexPlugin to preload multiple types.
This gives the ability to preload multiple tag types and override duplicate injections.

I was one of the original people to request the ability to automate preloading resources within the index.html via link tags. I noticed that y'all implemented it for script files but the request was for CSS files so I decided to take a stab at adding it in for you. I did this within the raw editor on github so nothing has been tested but it should work and give you a sense of what it should be. 

Let me know what you think. Thanks!